cli/requirements.txt
urlpath==1.1.2
packaging==16.8
attrdict==2.0.0
requests==2.11.1
ruamel.yaml==0.12.15
